requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
plotly==5.18.0
pillow==10.2.0 
loguru==0.6.0
groq==0.4.2
openai==1.1.0
openpyxl==3.1.2
python-dotenv==1.0.1
httpx==0.24.1